Conflicts:
	firmware/openos/openwsn/02a-MAClow/IEEE802154E.c
	firmware/openos/openwsn/openwsn.h
	firmware/openos/projects/common/03oos_macpong/03oos_macpong.c

+++ firmware/openos/drivers/common/openserial.h
@@ -63,7 +63,7 @@ enum {
 //=========================== prototypes ======================================
 
 void    openserial_init();
-error_t openserial_printStatus(uint8_t statusElement, uint8_t* buffer, uint16_t length);
+error_t openserial_printStatus(uint8_t statusElement, uint8_t* buffer, uint8_t length);
 error_t openserial_printInfo(uint8_t calling_component, uint8_t error_code,
                               errorparameter_t arg1,
                               errorparameter_t arg2);
